+++ client/.vscode/settings.json
@@ -1,0 +1,7 @@
+{
+    "gopls": {
+        "build.buildFlags": [
+            "-tags=linux 386 darwin !cgo"
+        ]
+    }
+}
